# docs/swagger.yaml
basePath: /
definitions:
...
  category.NewCategory:
    properties:
      name:
        maxLength: 50
        type: string
      parent_id:
        minimum: 0
        type: integer
    required:
    - name
    type: object
  developer.ChangeRole:
    properties:
      role:
        type: string
      user_id:
        type: integer
    type: object
...
info:
  contact: {}
  title: Swagger Example API
  version: "1.0"
paths:
...
  /category/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
      security:
      - BearerAuth: []
      summary: Delete Category By Id
      tags:
      - category
    get:
      consumes:
      - application/json
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
      summary: Get Category By Id
      tags:
      - category
  /category/{id}/{name}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: New Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
      security:
      - BearerAuth: []
      summary: Edit Category By Id
      tags:
      - category
  /category/add:
    post:
      consumes:
      - application/json
      description: Adding new category by admin
      parameters:
      - description: Adding new category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/category.NewCategory'
      produces:
      - application/json
      responses:
        "201":
          description: Created
      security:
      - BearerAuth: []
      summary: Add Category
      tags:
      - category
...
  /developer/change-role:
    put:
      consumes:
      - application/json
      parameters:
      - description: Changing user role
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/developer.ChangeRole'
      produces:
      - application/json
      responses: {}
      summary: change user role
      tags:
      - developer
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"